+++ app/controllers/accounts_controller.rb
@@ -1,46 +1,60 @@
 class AccountsController < ApplicationController
   def create
     params.permit!
-    result = check_request_body params
-    if result.empty?
-      account = Account.new
-      result, obj = account.create params
-      if result
-        render :status => :created, :json => obj
+
+    if params[:accounts]
+      required_column_names = [:account_type, :date, :content, :category, :price]
+      columns = params[:accounts].slice *required_column_names
+      absent_columns = [].tap do |arr|
+        required_column_names.each do |key|
+          arr << key unless columns[key]
+        end
+      end
+      if absent_columns.empty?
+        errors = [].tap do |array|
+          array << {:error_code => 'invalid_value_account_type'} unless columns[:account_type] =~ /income|expense/
+          array << {:error_code => 'invalid_value_date'} unless columns[:date] =~ /\A\d{4}-\d{2}-\d{2}\z/
+          array << {:error_code => 'invalid_value_price'} unless columns[:price] =~ /\A[1-9]\d*\z/
+        end
+        if errors.empty?
+          begin
+            account = Account.create!(columns)
+            render :status => :created, :json => account
+          rescue ActiveRecord::RecordInvalid => e
+            errors = e.record.errors.messages.keys.map do |column|
+              {:error_code => "invalid_value_#{column}"}
+            end
+            render :status => :bad_request, :json => errors
+          end
+        else
+          render :status => :bad_request, :json => errors
+        end
       else
-        errors = obj.map{|param| {:error_code => "invalid_value_#{param}"} }
-        render :status => :bad_request, :json => errors
+        absent_columns.map!{|column| {:error_code => "absent_param_#{column}"} }
+        render :status => :bad_request, :json => absent_columns
       end
     else
-      errors = result.map{|param| {:error_code => "absent_param_#{param}"} }
-      render :status => :bad_request, :json => errors
+      render :status => :bad_request, :json => [{:error_code => 'absent_param_accounts'}]
     end
   end
 
   def read
     params.permit!
-    result = check_request_body params
-    if result.empty?
-      account = Account.new
-      result, obj = account.show params
-      if result
-        render :status => :ok, :json => obj
-      else
-        errors = obj.map{|param| {:error_code => "invalid_value_#{param}"} }
-        render :status => :bad_request, :json => errors
-      end
+
+    result, obj = Account.show params
+    if result
+      render :status => :ok, :json => obj
     else
-      errors = result.map{|param| {:error_code => "absent_param_#{param}"} }
+      errors = obj.map{|param| {:error_code => "invalid_value_#{param}"} }
       render :status => :bad_request, :json => errors
     end
   end
 
   def update
     params.permit!
-    result = check_request_body params
-    if result.empty?
-      account = Account.new
-      result, obj = account.update params
+
+    if params[:with]
+      result, obj = Account.update params
       if result
         render :status => :ok, :json => obj
       else
@@ -48,68 +62,34 @@ class AccountsController < ApplicationController
         render :status => :bad_request, :json => errors        
       end
     else
-      errors = result.map{|param| {:error_code => "absent_param_#{param}"} }
-      render :status => :bad_request, :json => errors
+      render :status => :bad_request, :json => [{:error_code => 'absent_param_with'}]
     end
   end
 
   def delete
     params.permit!
-    result = check_request_body params
-    if result.empty?
-      account = Account.new
-      result, obj = account.destroy params
-      if result
-        render :status => :no_content, :nothing => true
-      else
-        errors = obj.map{|param| {:error_code => "invalid_value_#{param}"} }
-        render :status => :bad_request, :json => errors        
-      end
+
+    result, obj = Account.destroy params
+    if result
+      render :status => :no_content, :nothing => true
     else
-      errors = result.map{|param| {:error_code => "absent_param_#{param}"} }
-      render :status => :bad_request, :json => errors
+      errors = obj.map{|param| {:error_code => "invalid_value_#{param}"} }
+      render :status => :bad_request, :json => errors        
     end
   end
 
   def settle
     params.permit!
-    result = check_request_body params
-    if result.empty?
-      account = Account.new
-      result, obj = account.settle params[:period]
+
+    if params[:interval]
+      result, obj = Account.settle params[:interval]
       if result
         render :status => :ok, :json => obj
       else
-        errors = obj.map{|param| {:error_code => "invalid_value_#{param}"} }
-        render :status => :bad_request, :json => errors        
+        render :status => :bad_request, :json => [{:error_code => 'invalid_value_interval'}]
       end
     else
-      errors = result.map{|param| {:error_code => "absent_param_#{param}"} }
-      render :status => :bad_request, :json => errors
+      render :status => :bad_request, :json => [{:error_code => 'absent_param_interval'}]
     end
-  end
-
-  private
-
-  def check_request_body(request_params)
-    absent_params = []
-    case request_params[:action]
-    when 'create'
-      account = params[:accounts]
-      absent_params << :account_type unless account[:account_type]
-      absent_params << :date unless account[:date]
-      absent_params << :content unless account[:content]
-      absent_params << :category unless account[:category]
-      absent_params << :price unless account[:price]
-    when 'read'
-    when 'update'
-      absent_params << :with if not request_params[:with] or request_params[:with].empty?
-    when 'delete'
-    when 'settle'
-      absent_params << :period unless request_params[:period]
-    else
-      return false
-    end
-    absent_params
   end
 end