+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 algieba
 =======
 
-- 1.0.0: [NEW] create app
+- 1.0.1:
+  - [ADD] module and functional test
+  - [FIX] controller and model for account
+
+- 1.0.0:
+  - [NEW] create app